+++ examples/research_agent/question_generator.py
@@ -98,8 +98,8 @@ def create_question_generator_langchain_tool(
 
     assert isinstance(prompt, BasePromptTemplate), "Prompt must be a string or a BasePromptTemplate"
 
-    def partial_inserter(question: Question):
-        out = QuestionInsertionTool(graph=graph, question=question).to_langchain_tool()
+    def partial_inserter(input_dict: dict):
+        out = QuestionInsertionTool(graph=graph, question=input_dict["question"]).to_langchain_tool()
         return (out,)
 
     def insert_questions(input_dict) -> None:
@@ -119,11 +119,11 @@ def create_question_generator_langchain_tool(
         | RunnableLambda(print_passthrough)
         | {
             "subquestions": RunnablePassthrough.assign(question_text=lambda x: x["question"]["question"].question)
-            | RunnableLambda(print_passthrough)
             | prompt.partial(num_questions=max_questions)
             | llm,
             "question_inserter": RunnablePassthrough(),
         }
+        | RunnableLambda(print_passthrough)
         | RunnableLambda(insert_questions)
     )
